cloud_functions/requirements.txt
pytz>=2022.1
requests>=2.27.1
google-cloud-pubsub>=2.12.1